+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## master (unreleased)
 * Adds favicon build generator. ([@abhaynikam][])
+* Adds Pundit install generator. ([@CiTroNaK][https://github.com/CiTroNaK])
 
 ## 0.4.0 (October 23rd, 2020)
 * Adds CircleCI install generator. ([@abhaynikam][])